requirements.txt
bleach==6.2.0
blinker==1.8.2
Bootstrap-Flask==2.4.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==2.3.3
Flask-CKEditor==1.0.0
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
webencodings==0.5.1
Werkzeug==3.1.1
WTForms==3.2.1
